requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.26.2
torch==2.1.1
torchvision==0.16.1
pydantic==2.5.1
paddleocr==2.7.0
easyocr==1.7.1
transformers==4.35.2
ultralytics==8.0.208
